[refactor] #28 school id 타입 Long 형으로 변환!

+++ src/main/java/com/ggang/be/api/schoolMajor/service/SchoolMajorService.java
@@ -5,5 +5,5 @@ import java.util.List;
 
 public interface SchoolMajorService {
 
-    List<SearchSchoolMajorVo> findSchoolMajorBySchoolAndMajorName(Integer id, String schoolMajorKeyword);
+    List<SearchSchoolMajorVo> findSchoolMajorBySchoolAndMajorName(Long id, String schoolMajorKeyword);
 }
